requirements.txt
# Install dependencies using 'pip install -r requirements.txt'
rasterio
pandas
numpy
tensorflow
scikit-learn
PyYAML

